+++ src/kompe/cubed_sphere/__init__.py
@@ -2,6 +2,7 @@
 
 from kompe.cubed_sphere.cs_basis import GlobalCSBasis
 from kompe.cubed_sphere.cs_grid import GlobalCSMesh
+from kompe.cubed_sphere.diffutils import lcm_arr, stencil
 from kompe.cubed_sphere.plot import RegionalCSPlot
 from kompe.cubed_sphere.regional import (
     REGIONAL_CS_GRID_SCHEMA,
@@ -11,15 +12,31 @@ from kompe.cubed_sphere.regional import (
     RegionalCSOperators,
     RegionalCSProjection,
 )
+from kompe.cubed_sphere.spherical import (
+    car_to_sph,
+    ecef_to_enu,
+    enu_to_ecef,
+    sph_to_car,
+    sph_to_sph,
+    tangent_vector,
+)
 
 __all__ = [
-    "GlobalCSBasis",
-    "GlobalCSMesh",
     "REGIONAL_CS_GRID_SCHEMA",
     "REGIONAL_CS_GRID_SCHEMA_VERSION",
+    "GlobalCSBasis",
+    "GlobalCSMesh",
     "RegionalCSGrid",
     "RegionalCSGridSpec",
     "RegionalCSOperators",
-    "RegionalCSProjection",
     "RegionalCSPlot",
+    "RegionalCSProjection",
+    "car_to_sph",
+    "ecef_to_enu",
+    "enu_to_ecef",
+    "lcm_arr",
+    "sph_to_car",
+    "sph_to_sph",
+    "stencil",
+    "tangent_vector",
 ]